+++ incubating/argo-cd-sync/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.4.3] - 2024-02-22
+### Fixed
+intercepting application not found for better error message
+
+### Changed
+Move the creation of the link to the application earlier
+Exit with error when app is in OUT_OF_SYNC state
+
 ## [1.4.2] - 2024-01-17
 ### Changed
 New graphql call to speed up query